move typedef Complex and single_qubit_error out from stationary qubit to hardware monitor
